Import the Magic Comprehensive Rules (as of 2020.04.17)
The Magic Comprehensive Rules document is not part of Magicum. It is
reproduced here only as a reference.

Link:
https://magic.wizards.com/en/game-info/gameplay/rules-and-formats/rules

Original
https://media.wizards.com/2020/downloads/MagicCompRules%2020200417.txt

There was an encoding problem when opening the file in Emacs (when saved
with Firefox).

% file magic-comprehensive-rules--2020.04.17.txt
magic-comprehensive-rules--2020.04.17.txt: UTF-8 Unicode text, with very
long lines, with CRLF, CR line terminators

In Emacs, line endings where ignored, displaying "^M" in its place. The
file was correctly displayed in gedit.

Workaround: select all text in Firefox, copy it to gedit and save it as
a new file.

% file magic-comprehensive-rules--2020.04.17.txt
magic-comprehensive-rules--2020.04.17.txt: UTF-8 Unicode text, with very
long lines