specs: hack: make test work
This "fix" is, in fact, an ugly hack - which does not scale.

Problem: it is very cumbersome to reuse the specs when they are required
from another namespace. This is because the specs' fully qualified names
are being autoresolved (via `::`) to the namespace where they are
defined.

Example: see how complect it is to call the specs from
`magicum.specs.game` in `magicum.specs.game-test`.

TODO: They should be defined in a way that eases their use...